[init] mousecase factory style
